added liquidations

+++ analysis/get_healthfactor_drop_before_liquidation.py
@@ -8,6 +8,7 @@ sys.path.append("../scripts/")
 from utils import get_proxy_address, get_cached_abi
 from web3.providers.rpc import HTTPProvider
 import csv
+import os
 
 # Connect to an Ethereum Node
 api_url = "http://localhost:8545"  # your node url here
@@ -31,10 +32,13 @@ borrow_df.sort_values(by='block_timestamp_unix', inplace=True)
 borrow_df.reset_index(drop=True, inplace=True)
 
 # Add new columns for health factor block and blocks until liquidation if not already present
-if 'block_number_when_healthfactor_above_one' not in liquidations_df.columns:
-	liquidations_df['block_number_when_healthfactor_above_one'] = np.nan
-if 'blocks_for_liquidation' not in liquidations_df.columns:
-	liquidations_df['blocks_for_liquidation'] = np.nan
+new_cols = ['block_number_when_healthfactor_above_one','blocks_for_liquidation']
+added_cols = list(set(new_cols).difference(liquidations_df.columns))
+
+if len(added_cols) > 0:
+	for c in added_cols:
+		liquidations_df[c] = -1
+	liquidations_df = liquidations_df.astype( { c: int for c in added_cols } )
 
 # Path to output and error files
 output_file = "../data/liquidations_health_factor_details.csv"
@@ -64,13 +68,25 @@ def get_first_borrow_block_number(user):
 	else:
 		return None  # Return None if the user doesn't exist in the DataFrame
 
-headers = list(liquidations_df.columns) + ['block_number_when_healthfactor_above_one','blocks_for_liquidation']
-with open(partial_output_file, 'w', newline='') as f:
-	writer = csv.DictWriter(f, fieldnames=headers)
-	writer.writeheader()
+# Setup partial results file
+headers = list(liquidations_df.columns) + added_cols
+if not os.path.exists(partial_output_file):
+	with open(partial_output_file, 'w', newline='') as f:
+		writer = csv.DictWriter(f, fieldnames=headers)
+		writer.writeheader()
+	completed_rows = set()
+else:
+	partial_df = pd.read_csv( partial_output_file )
+	completed_rows = set( partial_df.transactionHash.unique() )
+
+unfinished_df = liquidations_df[ ~liquidations_df.transactionHash.isin(completed_rows) ]
+assert len( unfinished_df ) + len(completed_rows) == len(liquidations_df), 'Uncompleted + Completed != Total' 
 
 # Iterate over each row in the dataframe
 for i in tqdm(range(len(liquidations_df))):
+	if liquidations_df.transactionHash.iloc[i] in completed_rows:
+		continue
+
 	user = liquidations_df.user.iloc[i]
 	block_number = int(liquidations_df.blockNumber.iloc[i])
 	first_borrow_block = get_first_borrow_block_number(user)
